requirements.txt
Django==5.0.6
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-modeltranslation==0.18.11
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0
python-decouple==3.8
Pillow==10.3.0
django-extensions==3.2.3
